+++ common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 3.0.0 - 2025-09-05
+
+### Added (2)
+
+- Support automatic session re-logon on reconncetions/renewals when session is already logged on (`Session re-logon` option on `WebSocketAPIBase`).
+- Added the `api_key` parameter to include `apiKey` in WebsocketAPI request parameters.
+
+### Changed (2)
+
+- Fixed return type mismatch by returning the raw value.
+- Updated `WebsocketAPI` user data return value to match its type.
+
 ## 2.0.0 - 2025-08-22
 
 ### Changed (3)